add 2 custom Hooks
